# openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Microservice Upload File API",
    "version": "1.0.0",
    "description": "API for file upload and download operations"
  },
  "servers": [
    {
      "url": "http://localhost:9087",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root endpoint",
        "responses": {
          "200": {
            "description": "Welcome message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Hello World"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check endpoint",
        "responses": {
          "200": {
            "description": "Health status of the application",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "appName": {
                      "type": "string",
                      "example": "Upload file"
                    },
                    "version": {
                      "type": "string",
                      "example": "1.0.0"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/echo": {
      "post": {
        "summary": "Echo endpoint",
        "description": "Returns the JSON body sent in the request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Echoed response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/upload": {
      "post": {
        "summary": "File upload endpoint",
        "description": "Upload a file using multipart/form-data",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful upload response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "filename": {
                      "type": "string",
                      "example": "1693741852123_example.txt"
                    },
                    "size": {
                      "type": "number",
                      "example": 12345
                    },
                    "type": {
                      "type": "string",
                      "example": "text/plain"
                    },
                    "path": {
                      "type": "string",
                      "example": "/tmp/uploads/1693741852123_example.txt"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Content-Type must be multipart/form-data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/download/{filename}": {
      "get": {
        "summary": "File download endpoint",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of the file to download"
          }
        ],
        "responses": {
          "200": {
            "description": "File content",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  }
}